requirements.txt

gradio
huggingface_hub
PyMuPDF
sentence-transformers
faiss-cpu
numpy
pydantic
hf-transfer
protobuf
click

